Finish removing the CreateWalletReducer
