+++ src/actions/CreateWalletActions.tsx
@@ -5,10 +5,10 @@ import { Alert } from 'react-native'
 import { sprintf } from 'sprintf-js'
 
 import { ButtonsModal } from '../components/modals/ButtonsModal'
+import { AccountActivationPaymentInfo, HandleActivationInfo } from '../components/scenes/CreateWalletAccountSelectScene'
 import { Airship } from '../components/services/AirshipInstance'
 import { SPECIAL_CURRENCY_INFO } from '../constants/WalletAndCurrencyConstants'
 import { lstrings } from '../locales/strings'
-import { AccountActivationPaymentInfo, HandleActivationInfo } from '../reducers/scenes/CreateWalletReducer'
 import { getExchangeDenomByCurrencyCode } from '../selectors/DenominationSelectors'
 import { TokenWalletCreateItem } from '../selectors/getCreateWalletList'
 import { config } from '../theme/appConfig'
