+++ src/reducers/uiReducer.ts
@@ -5,13 +5,11 @@ import { Action } from '../types/reduxTypes'
 import { exchangeInfo } from './ExchangeInfoReducer'
 import { fio, FioState } from './FioReducer'
 import { passwordReminder, PasswordReminderState } from './PasswordReminderReducer'
-import { createWallet, CreateWalletState } from './scenes/CreateWalletReducer'
 import { fioAddress, FioAddressSceneState } from './scenes/FioAddressSceneReducer'
 import { settings, SettingsState } from './scenes/SettingsReducer'
 import { wallets, WalletsState } from './scenes/WalletsReducer'
 
 export interface UiState {
-  readonly createWallet: CreateWalletState
   readonly exchangeInfo: ExchangeInfo
   readonly fio: FioState
   readonly fioAddress: FioAddressSceneState
@@ -22,7 +20,6 @@ export interface UiState {
 }
 
 const uiInner = combineReducers<UiState, Action>({
-  createWallet,
   exchangeInfo,
   fio,
   fioAddress,
